Simple port of Class2Go's chat feature
This is intended to be a just-for-fun hack to get to know the edX
codebase a bit and see how to integrate a new feature. To do this
right, we'll probably want to turn this into an XBlock.

For now, this just points at a custom channel we've created on our
Jabber server.